requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
xgboost==2.1.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
imbalanced-learn==0.12.3
joblib==1.4.2
kafka-python==2.0.2
pydantic==2.9.2
prometheus-client==0.21.0
structlog==24.1.0
pytest==8.3.2
httpx==0.27.2
